pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "card-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "requests",
    "a2a-sdk[all]",
    "httpx",
    "uvicorn",
    "starlette",
    "sentence-transformers>=2.7.0",
    "torch>=2.0.0",
    "numpy",
    "accelerate>=1.12.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rich>=14.3.3",
    "google-genai>=1.64.0",
    "zai-sdk>=0.2.2",
    "jinja2",
    "datasets>=4.5.0",
    "huggingface-hub>=0.34.0",
    "hf-transfer>=0.1.9",
    "pydantic>=2.12.5",
    "demucs>=4.0.1",
    "faster-whisper>=1.1.1",
    "nemo-toolkit[asr]>=2.4.0",
    "torchaudio>=2.0.0",
    "soundfile>=0.13.1",
    "torchcodec>=0.10.0",
    "nltk>=3.8.1",
    "ctc-forced-aligner",
    "deepmultilingualpunctuation",
    "tiktoken",
    "textual",
    "unidecode"
]

[project.scripts]
card-framework-main = "card_framework.cli.main:main"
card-framework-setup = "card_framework.cli.setup_and_run:main"
card-framework-calibrate = "card_framework.cli.calibrate:main"
card-framework-eval = "card_framework.cli.eval:main"
card-framework-ui = "card_framework.cli.ui_cli:main"
card-framework-benchmark = "card_framework.benchmark.run:main"
card-framework-qa = "card_framework.benchmark.qa:main"
card-framework-diarization = "card_framework.benchmark.diarization:main"
card-framework-summary-matrix = "card_framework.cli.run_summary_matrix:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
card_framework = [
    "config/*.yaml",
    "prompts/templates/*.jinja2",
    "benchmark/*.yaml",
    "benchmark/*.md",
    "benchmark/manifests/*.json",
    "benchmark/rubrics/*.json",
    "audio_pipeline/gateways/*.yaml",
    "_vendor/index_tts/pyproject.toml",
    "_vendor/index_tts/uv.lock",
    "_vendor/index_tts/**/*.py",
    "_vendor/index_tts/**/*.json",
    "_vendor/index_tts/**/*.jsonl",
    "_vendor/index_tts/**/*.md",
    "_vendor/index_tts/**/*.model",
    "_vendor/index_tts/**/*.txt",
    "_vendor/index_tts/**/*.yaml",
    "_vendor/index_tts/**/*.yml",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
# Optional fallback index kept for environments that need CUDA 12.6 wheels.
# To switch, temporarily point `tool.uv.sources` torch/torchaudio to `pytorch-cu126`.

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }
ctc-forced-aligner = { git = "https://github.com/MahmoudAshraf97/ctc-forced-aligner.git" }
deepmultilingualpunctuation = { git = "https://github.com/oliverguhr/deepmultilingualpunctuation.git" }

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
xfail_strict = true
junit_family = "xunit2"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--durations=10",
    "--durations-min=1.0",
]
markers = [
    "unit: deterministic fast unit tests",
    "integration: component integration tests using local dependencies",
    "contract: shared interface/adapter behavior tests",
    "slow: intentionally slow tests",
    "gpu: tests that require GPU acceleration",
    "network: tests that require external network access",
]
